Move onNewCachedPsk down to FizzClientHandshake
